fix(directory): Westlake Pointe rebrand — "Brookdale Gardens at Westlake" is stale
The dry-run of the address fix matched 27569 Detroit Rd to "Westlake Pointe Senior
Living" (4.4★/35), not Brookdale Gardens. Confirmed: the building was REBRANDED —
directories list it as "Westlake Pointe, formerly Brookdale Gardens at Westlake".
Our row also stored the wrong street (Westlake Village's), which caused the shared
Google rating.

Supersedes fix-brookdale-gardens-address.ts (removed) with a full rebrand fix:
name → Westlake Pointe Senior Living, address → 27569 Detroit Rd (+re-geocode),
rating → the Westlake Pointe Google place, description de-staled (old brand token
replaced). Guarded: writes only when the Places match is confidently the
Detroit-Rd "Westlake Pointe" place (name + 27569 + a different place id than
Westlake Village); otherwise aborts. NOT a dedup; dry-run default.
